requirements.txt
langgraph
langchain
langchain-community
streamlit
python-dotenv
yfinance
wikipedia
requests
beautifulsoup4
pandas